pyproject.toml
[project]
name = "plataforma-seal"
version = "0.1.0"
description = "Plataforma SEAL - Sistema de Estrategistas de Alta Performance"
authors = [
    {name = "danielmarques-qv"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "django-ninja>=1.3,<2.0",
    "psycopg2-binary>=2.9,<3.0",
    "python-dotenv>=1.0,<2.0",
    "pyjwt>=2.8,<3.0",
    "cryptography>=41.0,<43.0",
    "django-cors-headers>=4.3,<5.0",
    "httpx>=0.27,<1.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
seal = "manage:main"

[tool.poetry]
package-mode = false